updated units to treat unknown/None more reasonably fixed serious errors in baseline allowed SpectrosopicAxis input when declaring a class voigtfitter now has components don't know what I fixed in plotters.... setup.py probably matches Jordan's, as does config_default

+++ setup.py
@@ -21,4 +21,5 @@ CWD = os.getcwd()
 if not os.path.exists('%s/.pyspeckit' % HOME):
     os.mkdir('%s/.pyspeckit' % HOME)
     
-shutil.copyfile('%s/pyspeckit/config_default' % CWD, '%s/.pyspeckit/config' % HOME)
+if not os.path.exists('%s/.pyspeckit/config' % HOME):
+    shutil.copyfile('%s/pyspeckit/config_default' % CWD, '%s/.pyspeckit/config' % HOME)
